+++ src/app/commands/bounty/Bounty.ts
@@ -67,6 +67,12 @@ module.exports = class Bounty extends SlashCommand {
 							description: 'What is the reward? (i.e 100 BANK)',
 							required: true,
 						},
+						{
+							name: 'copies',
+							type: CommandOptionType.INTEGER,
+							description: 'How many bounties should be published? (level 3+, max 100)',
+							required: false,
+						},
 					],
 				},
 				{
@@ -248,6 +254,7 @@ module.exports = class Bounty extends SlashCommand {
 	
 	buildBountyCreateNewParams(ctxOptions): BountyCreateNew {
 		const [reward, symbol] = (ctxOptions.reward != null) ? ctxOptions.reward.split(' ') : [null, null];
+		const copies = (ctxOptions.copies == null || ctxOptions.copies <= 0) ? 1 : ctxOptions.copies;
 		let scale = reward.split('.')[1]?.length;
 		scale = (scale != null) ? scale : 0;
 		return {
@@ -257,6 +264,7 @@ module.exports = class Bounty extends SlashCommand {
 				currencySymbol: symbol,
 				scale: scale,
 			},
+			copies: copies,
 		};
 	}
 };